refactor: apply 임시저장 분리
